+++ c-ares/src/lib.rs
@@ -37,6 +37,7 @@
 mod macros;
 mod a;
 mod aaaa;
+mod addrinfo;
 mod caa;
 mod channel;
 mod cname;
@@ -67,6 +68,10 @@ mod utils;
 // Re-export public interfaces.
 pub use crate::a::{AResult, AResults, AResultsIter};
 pub use crate::aaaa::{AAAAResult, AAAAResults, AAAAResultsIter};
+pub use crate::addrinfo::{
+    AddrInfoCName, AddrInfoCNameIter, AddrInfoFlags, AddrInfoHints, AddrInfoNode, AddrInfoNodeIter,
+    AddrInfoResults,
+};
 pub use crate::caa::{CAAResult, CAAResults, CAAResultsIter};
 #[cfg(cares1_29)]
 pub use crate::channel::ServerFailoverOptions;